refactor(ui): extract custom drawers for widget styling
Relocate drawers in their own module to make it easier to extend the system.

Key changes:
- Created a `drawers` module containing specialized drawer classes for:
  - Buttons (QPushButton)
  - Checkboxes (QCheckBox)
  - ComboBoxes (QComboBox)
  - Frames (QFrame)
  - Item Views (QStyledItemDelegate, QTreeView, QTableView)
  - Labels (QLabel)
  - Line Edits (QLineEdit)
  - Scroll Areas (QScrollArea)
  - Scroll Bars (QScrollBar)
  - Table Headers (QHeaderView)
  - Tooltips (QToolTip)
- Added `_utils.py` within the drawers module for shared logic like
  enum-to-string conversion and font styling.
- Refactored `AYONStyle` to integrate the new drawer registration
  mechanism.
- Added new widget variants for `QScrollArea` and `QFrame`.
