+++ client/ayon_ui_qt/variants.py
@@ -52,6 +52,10 @@ class QScrollBarVariants(Enum):
     Default = "default"
 
 
+class QScrollAreaVariants(Enum):
+    Default = "default"
+
+
 class QFrameVariants(Enum):
     Default = "default"
     Low = "low"
